requirements.txt
openai
flask
flask-cors
python-dotenv
python-pptx
reportlab
